feat: modularize mdx pages and blog posts
