+++ content/blog/hello-world/index.mdx
@@ -6,7 +6,7 @@ tags: ["WebDev", "Test"]
 description: First blog post, testing out MDX
 ---
 
-This is my first post on <Link to="/blog">luc.li/blog</Link>. Just testing stuff out.
+This is my first post on <InternalLink to="/blog">luc.li/blog</InternalLink>. Just testing stuff out.
 
 ## Why Gatsby?
